Support custom type prefix on root field groups
This allows users to override the type prefix of root field groups, instead of prefixing it with the post type by default.

For example, given the field group "Foo" that is used on post types "Product" and "Order", by default the following types are generated: `Product_Foo` and `Order_Foo`.

By using the custom prefix "ACF", only a single type is generated: `ACF_Foo`.

This should enable users to define reusable GraphQL Fragments for their field groups, regardless of which post type they are querying.